requirements.txt
pylint
pytest-repeat
pyyaml
flask
pydantic
pysftp

